add toolbar, hide_onscreen, hide_others, invert
old button model was not very user friendly. So we use a tool bar with tooltips explaining what they do.
while we are at it, add some new hide/show usefull shortcuts

TraitsUI's toolbar model is not very simple to use.
The new handler will try to find the action in the tcWindow class and in various childrens, accoring to where the function is easier to implement.