overview_cache: optimization to reduce the drawing time when lot of event
When looking at huge trace, the drawing is very slow
because a lot of rectangle have to be drawn.

We introduce an overview cache that merges events that are in the
same drawing area, so that it will only generate one rectangle instead of
multiple.

In that case, we also simplify the drawing by not drawing the comments
and the colors.

Signed-off-by: Pierre Tardy <tardyp@gmail.com>